requirements.txt
tensorflow==2.16.2
tensorflow-macos==2.16.2
tensorflow-metal==1.1.0
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn>=1.0.0 
Pillow>=8.0.0
